+++ src/RedisProxy.php
@@ -74,9 +74,9 @@ class RedisProxy
         $this->driversOrder = $this->supportedDrivers;
     }
 
-    public function setSentinelConnectionPool(array $sentinels, string $clusterId, int $database, float $timeout = 0.0, ?int $retryWait = null, ?int $maxFails = null)
+    public function setSentinelConnectionPool(array $sentinels, string $clusterId, int $database, float $timeout = 0.0, ?int $retryWait = null, ?int $maxFails = null, bool $writeToReplicas = true)
     {
-        $this->connectionPoolFactory = new SentinelConnectionPoolFactory($sentinels, $clusterId, $database, $timeout, $retryWait, $maxFails);
+        $this->connectionPoolFactory = new SentinelConnectionPoolFactory($sentinels, $clusterId, $database, $timeout, $retryWait, $maxFails, $writeToReplicas);
     }
 
     /**